refactor(mvm-build): migrate stage_job_dir to BuilderVmRuntime
Plan 97 Phase C PR-B-migrate commit 2 — first concern lifted out
of LibkrunBuilderVm into the hypervisor-agnostic helper.

Moves three items from libkrun_builder.rs to builder_vm_runtime.rs:

  - `stage_job_dir(job_dir, job)` — staging logic for cmd.sh
    (Flake jobs) and install_spec.json (Install jobs). ~150
    lines of cmd.sh body + 40 lines of dispatch.
  - `shell_single_quote_escape(s)` — POSIX shell quote-escape
    helper that the cmd.sh body uses to interpolate
    user-supplied flake refs / attr paths safely. Promoted to
    pub so external callers building per-job scripts (e.g.
    LibkrunBuilderVm::run_shell_script's validator) can reuse
    the same escape rules.
  - `INSTALL_SPEC_FILENAME` constant — the filename
    `mvm-builder-init` probes for to dispatch through the app-
    deps install pipeline (Plan 73 Followup B.2). Now pub.

`LibkrunBuilderVm::run_build` calls the runtime helper version
unchanged; libkrun_builder.rs imports the migrated symbols at
the top. No behavior change — the cmd.sh body is byte-identical
to the previous output (the only edit is replacing "Plan 72 W4"
in the cmd.sh header comment with a "Plan 97 Phase C migration"
note).

`stage_shell_job_dir` stays in libkrun_builder.rs for now — it's
used only by `run_shell_script` and is a simpler `fs::write`
that doesn't need the same multi-backend reuse story; Vz's
builder path won't need a separate shell-script staging hook.

All 307 mvm-build tests (with --features builder-vm) pass,
including the new builder_vm_runtime tests added in commit 1.
